gui: add imglist view
